+++ src/js/page-background.js
@@ -1,1 +1,1 @@
-TrelloChecker.run(true);
+Adapter.run(true);